feat(setup.sh): not restarting server if nothing to do, except when option --restart-anyway is specified (added logging too)

+++ xmage/setup.sh
@@ -21,6 +21,16 @@ XMAGE_SYSTEMD_UNIT_TIMER_PATH="$XMAGE_SYSTEMD_UNIT_DIR/$XMAGE_SYSTEMD_UNIT_TIMER
 
 set -e
 
+if [ "$1" = '-h' ] || [ "$1" = '--help' ]; then
+	echo "USAGE: $(basename "$0") [--restart-anyway]"
+	exit 0
+fi
+
+need_restart=false
+if [ "$1" = '--restart-anyway' ]; then
+	need_restart=true
+fi
+
 if ! id xmage >/dev/null 2>&1; then
 	echo "Error: xmage system user doesn't exist" >&2
 	echo "Please create it with the following command:" >&2
@@ -35,38 +45,46 @@ if [ ! -e "/var/lib/systemd/linger/xmage" ]; then
 	exit 4
 fi
 
-echo "Downloading web page of $XMAGE_DOMAIN ..."
+_log() {
+	if [ "$LOG_MODE" = 'datetime' ]; then
+		echo "$(date '+%Y-%m-%d %H:%M:%S')  $1"
+	elif [ "$LOG_MODE" != 'null' ]; then
+		echo "$1"
+	fi
+}
+
+_log "Downloading web page of $XMAGE_DOMAIN ..."
 dl_webpage_path="/tmp/webpage-from--$XMAGE_DOMAIN.html"
 wget -q -O "$dl_webpage_path" "$XMAGE_WEBSITE"
-echo "Downloaded to: $dl_webpage_path"
+_log "Downloaded to: $dl_webpage_path"
 
-echo "Searching for the app archive URL in the web page ..."
+_log "Searching for the app archive URL in the web page ..."
 xmage_archive_url_rel="$(grep '<a [^>]* href="files/mage-full_[^">]\+.zip">Download full BETA client</a>' "$dl_webpage_path" | sed 's#^.*href="\(files/mage-full_[^">]\+.zip\)".*$#\1#g')"
 if [ "$xmage_archive_url_rel" = '' ]; then
 	echo "Error: failed to find xmage archive URL in downloaded web page '$dl_webpage_path'" >&2
 	exit 2
 fi
-echo "Found: '$xmage_archive_url_rel'"
+_log "Found: '$xmage_archive_url_rel'"
 xmage_archive_url="$XMAGE_WEBSITE/$xmage_archive_url_rel"
-echo "Archive URL: '$xmage_archive_url'"
+_log "Archive URL: '$xmage_archive_url'"
 
-echo "Defining xmage archive filename ..."
+_log "Defining xmage archive filename ..."
 xmage_archive_filename="$(basename "$xmage_archive_url_rel")"
-echo "Archive filename: '$xmage_archive_filename'"
+_log "Archive filename: '$xmage_archive_filename'"
 
 if ! echo "$xmage_archive_filename" | grep -q '^mage-full_\([^_]\+\)_[0-9-]\{10\}_[0-9-]\{5\}\.zip$'; then
 	echo "Error: the archive filename '$xmage_archive_filename' has incorrect format" >&2
 	exit 3
 fi
 
-echo "Defining xmage latest version ..."
+_log "Defining xmage latest version ..."
 xmage_latest_version="$(echo "$xmage_archive_filename" | sed 's/^mage-full_\([^ ]\+\)\.zip$/\1/g')"
 xmage_latest_version_major="$(echo "$xmage_latest_version" | sed 's/^\([^_]\+\)_[0-9-]\{10\}_[0-9-]\{5\}$/\1/g')"
 xmage_latest_version_date="$(echo "$xmage_latest_version" | sed 's/^[^_]\+_\([0-9-]\{10\}_[0-9-]\{5\}\)$/\1/g;s/_//g')"
-echo "Latest version: $xmage_latest_version_major  $xmage_latest_version_date  ($xmage_latest_version)"
+_log "Latest version: $xmage_latest_version_major  $xmage_latest_version_date  ($xmage_latest_version)"
 
 if [ -L "$XMAGE_CURRENT_VERSION_PATH" ]; then
-	echo "Defining current installed version ..."
+	_log "Defining current installed version ..."
 	xmage_installed_version=none
 	xmage_installed_version_path="$(realpath "$XMAGE_CURRENT_VERSION_PATH")"
 	if [ -d "$xmage_installed_version_path" ]; then
@@ -74,20 +92,20 @@ if [ -L "$XMAGE_CURRENT_VERSION_PATH" ]; then
 		xmage_installed_version="$(echo "$xmage_installed_version_filename" | sed 's/^mage-full_\([^ ]\+\)$/\1/g')"
 
 		if ! echo "$xmage_installed_version" | grep -q '^[^_]\+_[0-9-]\{10\}_[0-9-]\{5\}$'; then
-			echo "Installed version: unknown  unknown  ($xmage_installed_version)"
+			_log "Installed version: unknown  unknown  ($xmage_installed_version)"
 			echo "Warning: the installed version filename '$xmage_installed_version' has incorrect format" >&2
 		else
 			xmage_installed_version_major="$(echo "$xmage_installed_version" | sed 's/^\([^_]\+\)_[0-9-]\{10\}_[0-9-]\{5\}$/\1/g')"
 			xmage_installed_version_date="$(echo "$xmage_installed_version" | sed 's/^[^_]\+_\([0-9-]\{10\}_[0-9-]\{5\}\)$/\1/g;s/_//g')"
-			echo "Installed version: $xmage_installed_version_major  $xmage_installed_version_date  ($xmage_installed_version)"
+			_log "Installed version: $xmage_installed_version_major  $xmage_installed_version_date  ($xmage_installed_version)"
 		fi
 	fi
 
-	echo "Analysing if version differs ..."
+	_log "Analysing if version differs ..."
 	if [ "$xmage_installed_version" != "$xmage_latest_version" ]; then
-		echo "Versions differs: will installed the latest one"
+		_log "Versions differs: will installed the latest one"
 	else
-		echo "Latest version is already installed and current: installing nothing"
+		_log "Latest version is already installed and current: installing nothing"
 		install_nothing=true
 	fi
 fi
@@ -96,33 +114,35 @@ if [ "$install_nothing" != 'true' ]; then
 
 	xmage_archive_dest="/tmp/$xmage_archive_filename"
 	if [ -f "$xmage_archive_dest" ]; then
-		echo "Using already downloaded archive '$xmage_archive_dest'"
+		_log "Using already downloaded archive '$xmage_archive_dest'"
 	else
-		echo "Downloading xmage archive ..."
+		_log "Downloading xmage archive ..."
 		wget -q -O "$xmage_archive_dest" "$xmage_archive_url"
 	fi
 
 	xmage_archive_dir="$(basename "$xmage_archive_filename" '.zip')"
         xmage_latest_version_dest_dir="$XMAGE_APPS_DIR/$xmage_archive_dir"
-	echo "Extracting xmage archive to '$xmage_latest_version_dest_dir' ..."
+	_log "Extracting xmage archive to '$xmage_latest_version_dest_dir' ..."
 	[ -d "$XMAGE_APPS_DIR" ] || mkdir "$XMAGE_APPS_DIR"
 	unzip -q -d "$xmage_latest_version_dest_dir" "$xmage_archive_dest" || true
 
-	echo "Creating/updating current version symlink ..."
+	_log "Creating/updating current version symlink ..."
 	rm -f "$XMAGE_CURRENT_VERSION_PATH"
 	ln -s "$xmage_archive_dir" "$XMAGE_CURRENT_VERSION_PATH"
 
-	echo "Updating xmage server configuration '$XMAGE_SERVER_CONFIG_PATH' ..."
+	_log "Updating xmage server configuration '$XMAGE_SERVER_CONFIG_PATH' ..."
 	sed 's/serverAddress="0\.0\.0\.0"/serverAddress="'"$HOST_DOMAIN"'"/g' -i "$XMAGE_SERVER_CONFIG_PATH"
 	sed 's/serverName="mage-server"/serverName="'"$HOST_DOMAIN"'"/g' -i "$XMAGE_SERVER_CONFIG_PATH"
 	sed 's/secondaryBindPort="-1"/secondaryBindPort="17172"/g' -i "$XMAGE_SERVER_CONFIG_PATH"
 	sed 's/leasePeriod="5000"/leasePeriod="'"$XMAGE_SERVER_LEASE_PERIOD"'"/g' -i "$XMAGE_SERVER_CONFIG_PATH"
 	sed 's/numAcceptThreads="2"/numAcceptThreads="'"$XMAGE_SERVER_THREADS"'"/g' -i "$XMAGE_SERVER_CONFIG_PATH"
+
+	need_restart=true
 fi
 
 
 if [ ! -f "$XMAGE_SERVER_START_SCRIPT" ]; then
-	echo "Installing xmage server start script ..."
+	_log "Installing xmage server start script ..."
 	cat >"$XMAGE_SERVER_START_SCRIPT" <<ENDCAT
 #!/bin/sh
 
@@ -151,7 +171,7 @@ export XDG_CONFIG_HOME
 installed_systemd_files=false
 [ -d "$XMAGE_SYSTEMD_UNIT_DIR" ] || mkdir -p "$XMAGE_SYSTEMD_UNIT_DIR"
 if [ ! -f "$XMAGE_SYSTEMD_UNIT_PATH" ]; then
-	echo "Installing xmage systemd unit '$XMAGE_SYSTEMD_UNIT_PATH' ..."
+	_log "Installing xmage systemd unit '$XMAGE_SYSTEMD_UNIT_PATH' ..."
 	cat >"$XMAGE_SYSTEMD_UNIT_PATH" <<ENDCAT
 [Unit]
 Description=XMage server
@@ -167,7 +187,7 @@ ENDCAT
 	installed_systemd_files=true
 fi
 if [ ! -f "$XMAGE_SYSTEMD_UNIT_TIMER_PATH" ]; then
-	echo "Installing xmage systemd unit timer '$XMAGE_SYSTEMD_UNIT_TIMER_PATH' ..."
+	_log "Installing xmage systemd unit timer '$XMAGE_SYSTEMD_UNIT_TIMER_PATH' ..."
 	cat >"$XMAGE_SYSTEMD_UNIT_TIMER_PATH" <<ENDCAT
 [Unit]
 Description=Run XMage server on boot
@@ -181,32 +201,38 @@ ENDCAT
 	installed_systemd_files=true
 fi
 if [ "$installed_systemd_files" = 'true' ]; then
-	echo "Reloading systemd daemon ..."
+	_log "Reloading systemd daemon ..."
 	systemctl --user daemon-reload
 
-	echo "Enabling systemd unit ..."
+	_log "Enabling systemd unit ..."
 	systemctl --user enable "$XMAGE_SYSTEMD_UNIT_FILENAME"
 
-	echo "Enabling systemd unit timer ..."
+	_log "Enabling systemd unit timer ..."
 	systemctl --user enable "$XMAGE_SYSTEMD_UNIT_TIMER_FILENAME"
+
+	need_restart=true
 fi
 
-echo "Stopping server ..."
-systemctl --user stop "$XMAGE_SYSTEMD_UNIT_FILENAME" || true
+if [ "$need_restart" = 'true' ]; then
+	_log "Stopping server ..."
+	systemctl --user stop "$XMAGE_SYSTEMD_UNIT_FILENAME" || true
 
-echo "Detecting a server running ..."
-server_pid="$(ps aux| (grep 'java .*-jar \./lib/mage-server-[^ ]\+\.jar' || true) | awk '{print $2}')"
-if [ "$server_pid" != '' ]; then
-	echo "Server is running with PID: '$server_pid'"
-	echo "Stopping currently running server ..."
-	kill "$server_pid"
-	if [ "$(ps -p "$server_pid" || true)" != '' ]; then
-		echo "Warning failed to kill running server with PID '$server_pid'" >&2
-		exit 4
+	_log "Detecting a server running ..."
+	server_pid="$(ps aux| (grep 'java .*-jar \./lib/mage-server-[^ ]\+\.jar' || true) | awk '{print $2}')"
+	if [ "$server_pid" != '' ]; then
+		_log "Server is running with PID: '$server_pid'"
+		_log "Stopping currently running server ..."
+		kill "$server_pid"
+		if [ "$(ps -p "$server_pid" || true)" != '' ]; then
+			echo "Warning failed to kill running server with PID '$server_pid'" >&2
+			exit 4
+		fi
+	else
+		_log "No server running"
 	fi
-else
-	echo "No server running"
-fi
 
-echo "Starting the server ..."
-systemctl --user start "$XMAGE_SYSTEMD_UNIT_FILENAME"
+	_log "Starting the server ..."
+	systemctl --user start "$XMAGE_SYSTEMD_UNIT_FILENAME"
+else
+	_log "Not restarting the server"
+fi
